+++ src/main/java/az/hrportal/hrportalapi/dto/employee/response/EmployeeSalaryResponseDto.java
@@ -5,8 +5,6 @@ import lombok.Getter;
 import lombok.Setter;
 import lombok.experimental.FieldDefaults;
 
-import java.math.BigDecimal;
-
 @Getter
 @Setter
 @FieldDefaults(level = AccessLevel.PRIVATE)
@@ -14,6 +12,6 @@ public class EmployeeSalaryResponseDto {
     Integer id;
     String fullName;
     String vacancyName;
-    BigDecimal netSalary;
-    BigDecimal grossSalary;
+    Float netSalary;
+    Float grossSalary;
 }
